+++ @navikt/core/tokens/darkside/tokens/colors/index.ts
@@ -1,0 +1,4 @@
+export {
+  globalColorDarkModeConfig,
+  globalColorLightModeConfig,
+} from "./global-config";
